Cleanup
Reduced data storage types.
Added Hotbar key switching.
Added an exit game key.
Removed mipmaps. (Redundant)
Reduced unnecessary computation for the block highlight.

+++ include/gui.h
@@ -1,14 +1,14 @@
 #define MAX_UI_ELEMENTS 3
 
 typedef struct {
-	float x;
-	float y;
-	float width;
-	float height;
-	int tex_x;
-	int tex_y;
-	int tex_width;
-	int tex_height;
+	uint16_t x;
+	uint16_t y;
+	uint16_t width;
+	uint16_t height;
+	uint8_t tex_x;
+	uint8_t tex_y;
+	uint8_t tex_width;
+	uint8_t tex_height;
 } ui_element_t;
 
 void init_ui();
